+++ src/test/java/com/example/server/service/MemberServiceTest.java
@@ -18,7 +18,7 @@ import org.springframework.mock.web.MockMultipartFile;
 
 import com.example.server.domain.image.model.Image;
 import com.example.server.domain.member.model.Member;
-import com.example.server.global.dto.MemberDto;
+import com.example.server.domain.member.api.dto.ProfileSaveRequest;
 import com.example.server.global.exception.BusinessException;
 import com.example.server.domain.member.persistence.MemberQueryRepository;
 import com.example.server.domain.member.application.MemberService;
@@ -43,14 +43,14 @@ public class MemberServiceTest {
 			.build();
 	}
 
-	private MemberDto.ProfileSaveRequest createProfileSaveRequest() {
-		return MemberDto.ProfileSaveRequest.builder()
+	private ProfileSaveRequest.ProfileSaveRequest createProfileSaveRequest() {
+		return ProfileSaveRequest.ProfileSaveRequest.builder()
 			.nickname("테스트닉네임")
 			.build();
 	}
 
-	private MemberDto.ProfileSaveRequest createBlankedProfileSaveRequest() {
-		return MemberDto.ProfileSaveRequest.builder()
+	private ProfileSaveRequest.ProfileSaveRequest createBlankedProfileSaveRequest() {
+		return ProfileSaveRequest.ProfileSaveRequest.builder()
 			.build();
 	}
 
@@ -64,7 +64,7 @@ public class MemberServiceTest {
 	@Test
 	public void passed_닉네임을_포함한_프로필_정보를_등록한다() {
 		Member member = createMember();
-		MemberDto.ProfileSaveRequest profileSaveRequest = createProfileSaveRequest();
+		ProfileSaveRequest.ProfileSaveRequest profileSaveRequest = createProfileSaveRequest();
 		when(memberQueryRepository.findByProviderId(member.getUsername())).thenReturn(Optional.of(member));
 		when(memberQueryRepository.existsByNickname(profileSaveRequest.getNickname())).thenReturn(false);
 
@@ -77,7 +77,7 @@ public class MemberServiceTest {
 	@Test
 	public void passed_요청필드가_비어있다면_수정하지_않는다() {
 		Member member = createMember();
-		MemberDto.ProfileSaveRequest blankedProfileSaveRequest = createBlankedProfileSaveRequest();
+		ProfileSaveRequest.ProfileSaveRequest blankedProfileSaveRequest = createBlankedProfileSaveRequest();
 		when(memberQueryRepository.findByProviderId(member.getUsername())).thenReturn(Optional.of(member));
 
 		memberService.updateProfile(blankedProfileSaveRequest, member.getUsername());
@@ -88,7 +88,7 @@ public class MemberServiceTest {
 
 	@Test
 	public void failed_회원을_찾을_수_없는_경우_예외가_발생한다() {
-		MemberDto.ProfileSaveRequest profileSaveRequest = createProfileSaveRequest();
+		ProfileSaveRequest.ProfileSaveRequest profileSaveRequest = createProfileSaveRequest();
 
 		assertThatThrownBy(() -> memberService.updateProfile(profileSaveRequest, "stranger"))
 			.isInstanceOf(BusinessException.class);
